requirements.txt
numpy>=1.26
pandas>=2.0
Pillow>=10.0
torch>=2.0
torchvision>=0.15
open-clip-torch>=2.24.0
gradio>=4.0
